Update redox
The script will check if certain tools like nmap, gobuster, whatweb, and jq are installed and will exit with an error message if any of them are not found. It will take one command-line argument which is the IP address to scan, and number of decoy IPs to run a scan on the target IP using nmap.

It's using various tools to gather information like Gobuster, WhatWeb and shodan to get information of target IP and then it is also doxing the IP address using an online service. The script then formats and organizes the information in a file named results.
It also using 'temp1', 'temp2' and 'temp3' as temporary files to store information before organizing in results file.